Module full_node::network_service

Background network service.

The NetworkService manages background tasks dedicated to connecting to other nodes. Importantly, its design is oriented towards the particular use case of the full node.

The NetworkService spawns one background task (using the Config::tasks_executor) for each active TCP socket, plus one for each TCP listening socket. Messages are exchanged between the service and these background tasks.

Structs

ChainConfig
Configuration for one chain.

Config
Configuration for a NetworkService.

Guarded
Fields of NetworkService behind a mutex.

NetworkService

Enums

Event
Event generated by the events reporters returned by NetworkService::new.

InitError
Error when initializing the network service.

Functions

connection_task
Asynchronous task managing a specific TCP connection.

multiaddr_to_socket
Builds a future that connects to the given multiaddress. Returns an error if the multiaddress protocols aren’t supported.