Module full_node::run

Modules

json_rpc_service

network_service

Background network service.

sync_service

Background synchronization service.

Functions

background_open_database

Since opening the database can take a long time, this utility function performs this operation in the background while showing a small progress bar to the user.

open_database

Opens the database from the filesystem, or create a new database if none is found.

run

Runs the node using the given configuration. Catches SIGINT signals and stops if one is detected.