Module full_node::run::sync_service

Background synchronization service.

The SyncService manages a background task dedicated to synchronizing the chain with the network. Importantly, its design is oriented towards the particular use case of the full node.

Structs

BlocksRequestId

Identifier for a blocks request to be performed.

Config

Configuration for a SyncService.

SyncBackground

SyncService

Background task that verifies blocks and emits requests.

SyncState

Summary of the state of the SyncService.

Enums

ToDatabase

Functions

start_database_write

Starts the task that writes blocks to the database.